Tidy up loading and add classify  (#121)
* Replace everything by missing if gdalfile throws error and rechunk the data

This will allow to read Sentinel-1 data even though one single time step is not available.

* Remove timestats.jl include

* Add classifier for RQA trend

With this we can save the RQA trend data in UInt8 data and can apply the threshold on that.

* Tidy up merge

* Add AWSS3 dependency

* Use Path consistently  in main

* Fix tests

* Remove AWSS3

This is actually not needed, because we are trying to keep the Path type and don't have to recast it.

* Add test for skipmissingmean

* Remove duplicate path definition in main

* Improve docstrings in rqatrend

* Fix main test with actual data

* Remove old unused loading path

* Remove aggregate_diskarray way as well

---------

Co-authored-by: Felix Cremer <felix.cremer@bgc-jena.mpg.de>